elevenedge/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
supabase==2.7.4
pydantic==2.9.2
pydantic-settings==2.5.2
openai-whisper==20240930
python-multipart==0.0.9
discord.py==2.4.0
requests==2.32.3
